* UserListAdapter list items view refactored * Typos fixed * Data dialog layout refactored * Function contains display fix * XML format refactored
Conflicts:
	catroid/res/values-de/strings.xml
	catroid/res/values/strings.xml

+++ catroid/src/org/catrobat/catroid/common/BrickValues.java
@@ -72,10 +72,10 @@ public final class BrickValues {
 
 	//Constants Lists
 	public static final double ADD_ITEM_TO_USERLIST = 1d;
-	public static final double DELETE_ITEM_OF_USERLIST = 1d;
-	public static final double INSERT_ITEM_INTO_USERLIST_INDEX = 1d;
+	public static final int DELETE_ITEM_OF_USERLIST = 1;
+	public static final int INSERT_ITEM_INTO_USERLIST_INDEX = 1;
 	public static final double INSERT_ITEM_INTO_USERLIST_VALUE = 1d;
-	public static final double REPLACE_ITEM_IN_USERLIST_INDEX = 1d;
+	public static final int REPLACE_ITEM_IN_USERLIST_INDEX = 1;
 	public static final double REPLACE_ITEM_IN_USERLIST_VALUE = 1d;
 
 	// Suppress default constructor for noninstantiability
